resolver: Compress all DNS answers
This patch makes mesos-dns compress all DNS answers. This improves
performance of many DNS questions which, when compressed, fit in 512
bytes. See https://tools.ietf.org/html/rfc1035#section-4.2.1

Fixes #170 because the native Go DNS resolver which is enabled when
statically compiling doesn't fallback to TCP when failing to decode
a UDP response larger than 512 bytes.

It isn't a perfect solution since an answer might be larger than 512
bytes even after compression but that's an issue with the client
resolver.