resolver: make setting of the Truncate bit configurable
Setting the Truncate bit in the DNS message when it is truncated is now configurable.

The default behaviour is to set the Truncate bit when appropriate.

When this option is set to false the Truncate bit will be unset in all DNS responses.

This is non-compliant behaviour but we support it for cases where clients are also non-compliant.